The audio player invalidates itself during idle rather than redrawing.
It was generating a paint event anyhow (at least on wxCocoa), causing a
second painting, and also this scary message to stderr on macOS:

"altrace_wx[38782:1137383] It does not make sense to draw an image when
[NSGraphicsContext currentContext] is nil.  This is a programming error.
Break on void _NSWarnForDrawingImageWithNoCurrentContext(void) to debug.
This will be logged only once.  This may break in the future."